pyproject.toml
# https://hackersandslackers.com/python-poetry-package-manager/
[tool.poetry]
name = "wws"
version = "0.1.1"
description = "A simple utility to synchronize local and remote paths"
authors = ["Daniel Porto <daniel.porto@gmail.com>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.7"
plumbum = "^1.6"
funcy = "^1.14"
pyyaml = "^5.3"
pprint = "^0.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
run = "wws:main"

[build-system]
requires = ["poetry>=1.0.9"]
build-backend = "poetry.masonry.api"

